refactor(surface): the public API is the product, not the parts that build it
